feat(forntend/backend)
Assign appoinment, fixes

+++ src/main/java/com/agendamiento/demo/Models/Dao/IUserDao.java
@@ -4,9 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.agendamiento.demo.Models.Entity.User;
 
-public interface IUserDao extends JpaRepository<User, Integer>{
+public interface IUserDao extends JpaRepository<User, Long>{
   
-
     public boolean existsByEmail(String email);
 
     public User findByEmail(String email);
